+++ bugsnag/tornado/__init__.py
@@ -3,7 +3,9 @@ from tornado.web import RequestHandler, HTTPError
 from tornado.wsgi import WSGIContainer
 from typing import Dict, Any
 from urllib.parse import parse_qs
-from bugsnag.utils import is_json_content_type
+from bugsnag.breadcrumbs import BreadcrumbType
+from bugsnag.utils import is_json_content_type, sanitize_url
+from bugsnag.legacy import _auto_leave_breadcrumb
 import bugsnag
 import json
 
@@ -74,9 +76,27 @@ class BugsnagRequestHandler(RequestHandler):
         middleware = bugsnag.configure().internal_middleware
         bugsnag.configure().runtime_versions['tornado'] = tornado.version
         middleware.before_notify(self.add_tornado_request_to_notification)
+        bugsnag.configure()._breadcrumbs.create_copy_for_context()
+
+        _auto_leave_breadcrumb(
+            'http request',
+            self._get_breadcrumb_metadata(),
+            BreadcrumbType.NAVIGATION
+        )
 
         if bugsnag.configuration.auto_capture_sessions:
             bugsnag.start_session()
+
+    def _get_breadcrumb_metadata(self) -> Dict[str, str]:
+        if not hasattr(self, 'request'):
+            return {}
+
+        metadata = {'to': self.request.path}
+
+        if 'Referer' in self.request.headers:
+            metadata['from'] = sanitize_url(self.request.headers['Referer'])
+
+        return metadata
 
     def _get_context(self):
         return "%s %s" % (self.request.method, self.request.uri.split("?")[0])