+++ tests/test_utils.py
@@ -1,15 +1,18 @@
 import unittest
 import json
 import timeit
-import sys
-import datetime
 import re
 import threading
 import uuid
+import logging
+import pytest
+from datetime import datetime, timedelta, timezone
 
 from bugsnag.utils import (SanitizingJSONEncoder, FilterDict,
                            is_json_content_type, parse_content_type,
-                           ThreadContextVar)
+                           ThreadContextVar, to_rfc3339, sanitize_url)
+
+logger = logging.getLogger(__name__)
 
 
 class TestUtils(unittest.TestCase):
@@ -19,8 +22,38 @@ class TestUtils(unittest.TestCase):
     def test_encode_filters(self):
         data = FilterDict({"credit_card": "123213213123", "password": "456",
                            "cake": True})
-        encoder = SanitizingJSONEncoder(keyword_filters=["credit_card",
-                                                         "password"])
+        encoder = SanitizingJSONEncoder(
+            logger,
+            keyword_filters=["credit_card", "password"]
+        )
+
+        sane_data = json.loads(encoder.encode(data))
+        self.assertEqual(sane_data, {"credit_card": "[FILTERED]",
+                                     "password": "[FILTERED]",
+                                     "cake": True})
+
+    def test_encode_filters_object_key(self):
+        object_key = object()
+        data = FilterDict({"password": "456", object_key: True})
+
+        encoder = SanitizingJSONEncoder(
+            logger,
+            keyword_filters=["password"]
+        )
+
+        actual = json.loads(encoder.encode(data))
+        expected = {"password": "[FILTERED]", str(object_key): True}
+
+        self.assertEqual(actual, expected)
+
+    def test_encode_filters_bytes(self):
+        data = FilterDict({b"credit_card": "123213213123", b"password": "456",
+                           "cake": True})
+        encoder = SanitizingJSONEncoder(
+            logger,
+            keyword_filters=["credit_card", "password"]
+        )
+
         sane_data = json.loads(encoder.encode(data))
         self.assertEqual(sane_data, {"credit_card": "[FILTERED]",
                                      "password": "[FILTERED]",
@@ -29,59 +62,68 @@ class TestUtils(unittest.TestCase):
     def test_sanitize_list(self):
         data = FilterDict({"list": ["carrots", "apples", "peas"],
                            "passwords": ["abc", "def"]})
-        encoder = SanitizingJSONEncoder(keyword_filters=["credit_card",
-                                                         "passwords"])
+
+        encoder = SanitizingJSONEncoder(
+            logger,
+            keyword_filters=["credit_card", "passwords"]
+        )
+
         sane_data = json.loads(encoder.encode(data))
         self.assertEqual(sane_data, {"list": ["carrots", "apples", "peas"],
                                      "passwords": "[FILTERED]"})
 
     def test_sanitize_valid_unicode_object(self):
         data = {"item": '\U0001f62c'}
-        encoder = SanitizingJSONEncoder(keyword_filters=[])
+        encoder = SanitizingJSONEncoder(logger, keyword_filters=[])
         sane_data = json.loads(encoder.encode(data))
         self.assertEqual(sane_data, data)
 
     def test_sanitize_nested_object_filters(self):
         data = FilterDict({"metadata": {"another_password": "My password"}})
-        encoder = SanitizingJSONEncoder(keyword_filters=["password"])
+
+        encoder = SanitizingJSONEncoder(
+            logger,
+            keyword_filters=["password"]
+        )
+
         sane_data = json.loads(encoder.encode(data))
         self.assertEqual(sane_data,
                          {"metadata": {"another_password": "[FILTERED]"}})
 
     def test_sanitize_bad_utf8_object(self):
         data = {"bad_utf8": "test \xe9"}
-        encoder = SanitizingJSONEncoder(keyword_filters=[])
+        encoder = SanitizingJSONEncoder(logger, keyword_filters=[])
         sane_data = json.loads(encoder.encode(data))
         self.assertEqual(sane_data, data)
 
     def test_sanitize_unencoded_object(self):
         data = {"exc": Exception()}
-        encoder = SanitizingJSONEncoder(keyword_filters=[])
+        encoder = SanitizingJSONEncoder(logger, keyword_filters=[])
         sane_data = json.loads(encoder.encode(data))
         self.assertEqual(sane_data, {"exc": ""})
 
     def test_json_encode(self):
         payload = {"a": "a" * 512 * 1024}
         expected = {"a": "a" * 1024}
-        encoder = SanitizingJSONEncoder(keyword_filters=[])
+        encoder = SanitizingJSONEncoder(logger, keyword_filters=[])
         self.assertEqual(json.loads(encoder.encode(payload)), expected)
 
     def test_filter_dict(self):
         data = FilterDict({"metadata": {"another_password": "My password"}})
-        encoder = SanitizingJSONEncoder(keyword_filters=["password"])
+        encoder = SanitizingJSONEncoder(logger, keyword_filters=["password"])
         sane_data = encoder.filter_string_values(data)
         self.assertEqual(sane_data,
                          {"metadata": {"another_password": "[FILTERED]"}})
 
     def test_decode_bytes(self):
         data = FilterDict({b"metadata": "value"})
-        encoder = SanitizingJSONEncoder(keyword_filters=["password"])
+        encoder = SanitizingJSONEncoder(logger, keyword_filters=["password"])
         sane_data = json.loads(encoder.encode(data))
         self.assertEqual(sane_data, {"metadata": "value"})
 
     def test_unfiltered_encode(self):
         data = {"metadata": {"another_password": "My password"}}
-        encoder = SanitizingJSONEncoder(keyword_filters=["password"])
+        encoder = SanitizingJSONEncoder(logger, keyword_filters=["password"])
         sane_data = json.loads(encoder.encode(data))
         self.assertEqual(sane_data, data)
 
@@ -144,7 +186,7 @@ class TestUtils(unittest.TestCase):
         """
         data = {"Test": ["a", "b", "c"]}
         data["Self"] = data
-        encoder = SanitizingJSONEncoder(keyword_filters=[])
+        encoder = SanitizingJSONEncoder(logger, keyword_filters=[])
         sane_data = json.loads(encoder.encode(data))
         self.assertEqual(sane_data,
                          {"Test": ["a", "b", "c"], "Self": "[RECURSIVE]"})
@@ -155,7 +197,7 @@ class TestUtils(unittest.TestCase):
         """
         data = {"Test": ["a", "b", "c"]}
         data["Self"] = data
-        encoder = SanitizingJSONEncoder(keyword_filters=[])
+        encoder = SanitizingJSONEncoder(logger, keyword_filters=[])
         sane_data = json.loads(encoder.encode(data))
         self.assertEqual(sane_data,
                          {"Test": ["a", "b", "c"], "Self": "[RECURSIVE]"})
@@ -168,7 +210,7 @@ class TestUtils(unittest.TestCase):
         Test that encoding the same object within a new object is not
         incorrectly marked as recursive
         """
-        encoder = SanitizingJSONEncoder(keyword_filters=[])
+        encoder = SanitizingJSONEncoder(logger, keyword_filters=[])
         data = {"Test": ["a", "b", "c"]}
         encoder.encode(data)
         data = {"Previous": data, "Other": 400}
@@ -184,7 +226,7 @@ class TestUtils(unittest.TestCase):
         """
         data = {"Test": ["a" * 128 * 1024, "b", "c"], "Other": {"a": 300}}
         data["Self"] = data
-        encoder = SanitizingJSONEncoder(keyword_filters=[])
+        encoder = SanitizingJSONEncoder(logger, keyword_filters=[])
         sane_data = json.loads(encoder.encode(data))
         self.assertEqual(sane_data,
                          {"Test": ["a" * 1024, "b", "c"],
@@ -197,25 +239,25 @@ class TestUtils(unittest.TestCase):
         """
         setup = """\
 import json
+import logging
 from tests.large_object import large_object_file_path
 from bugsnag.utils import SanitizingJSONEncoder
-encoder = SanitizingJSONEncoder(keyword_filters=[])
+
+logger = logging.getLogger(__name__)
+encoder = SanitizingJSONEncoder(logger, keyword_filters=[])
 with open(large_object_file_path()) as json_data:
     data = json.load(json_data)
         """
-        stmt = """\
-encoder.encode(data)
-        """
+
+        stmt = "encoder.encode(data)"
+
         time = timeit.timeit(stmt=stmt, setup=setup, number=1000)
         maximum_time = 6
-        if sys.version_info[0:2] <= (2, 6):
-            # json encoding is very slow on python 2.6 so we need to increase
-            # the allowable time when running on it
-            maximum_time = 18
-        self.assertTrue(time < maximum_time,
-                        "Encoding required {0}s (expected {1}s)".format(
-                            time, maximum_time
-                        ))
+
+        self.assertTrue(
+            time < maximum_time,
+            "Encoding required {0}s (expected {1}s)".format(time, maximum_time)
+        )
 
     def test_filter_string_values_list_handling(self):
         """
@@ -223,7 +265,7 @@ encoder.encode(data)
         parameter for backwards compatibility
         """
         data = {}
-        encoder = SanitizingJSONEncoder()
+        encoder = SanitizingJSONEncoder(logger)
         # no assert as we are just expecting this not to throw
         encoder.filter_string_values(data, ['password'])
 
@@ -233,7 +275,7 @@ encoder.encode(data)
         backwards compatibility
         """
         data = {}
-        encoder = SanitizingJSONEncoder()
+        encoder = SanitizingJSONEncoder(logger)
         # no assert as we are just expecting this not to throw
         encoder._sanitize(data, ['password'], ['password'])
 
@@ -243,7 +285,7 @@ encoder.encode(data)
         name or some other bad data is passed as a key in the payload
         dictionary.
         """
-        encoder = SanitizingJSONEncoder(keyword_filters=[])
+        encoder = SanitizingJSONEncoder(logger, keyword_filters=[])
 
         def foo():
             return "123"
@@ -253,7 +295,7 @@ encoder.encode(data)
                                  list(result.keys())[0]) is not None)
         self.assertEqual(list(result.values()), ["a"])
 
-        now = datetime.datetime.now()
+        now = datetime.now()
         result = json.loads(encoder.encode({now: "a"}))
         self.assertEqual(list(result.keys())[0], str(now))
         self.assertEqual(list(result.values()), ["a"])
@@ -284,7 +326,7 @@ encoder.encode(data)
                 }),
             }
         }
-        encoder = SanitizingJSONEncoder(keyword_filters=['password'])
+        encoder = SanitizingJSONEncoder(logger, keyword_filters=['password'])
         sane_data = json.loads(encoder.encode(data))
         self.assertEqual(sane_data, {
                             'level1-key1': {
@@ -331,7 +373,12 @@ encoder.encode(data)
                 },
             }
         })
-        encoder = SanitizingJSONEncoder(keyword_filters=['password', 'token'])
+
+        encoder = SanitizingJSONEncoder(
+            logger,
+            keyword_filters=['password', 'token']
+        )
+
         filtered_data = encoder.filter_string_values(data)
         self.assertEqual(filtered_data, {
                             'level1-key1': {
@@ -398,3 +445,83 @@ encoder.encode(data)
         self.assertFalse(is_json_content_type('text/plain'))
         self.assertFalse(is_json_content_type('json'))
         self.assertFalse(is_json_content_type('application/jsonfoo'))
+
+
+utc = timezone.utc
+plus_1 = timezone(timedelta(hours=1))
+plus_12 = timezone(timedelta(hours=12))
+plus_0_15 = timezone(timedelta(hours=0, minutes=15))
+plus_12_30 = timezone(timedelta(hours=12, minutes=30))
+minus_1 = timezone(timedelta(hours=-1))
+minus_12 = timezone(timedelta(hours=-12))
+minus_0_15 = timezone(timedelta(hours=0, minutes=-15))
+minus_12_30 = timezone(timedelta(hours=-12, minutes=-30))
+
+
+@pytest.mark.parametrize("dt, expected", [
+    (datetime(1, 1, 1, 1, 1, 1, 1), '0001-01-01T01:01:01.000'),  # noqa: E501
+    (datetime(1900, 1, 2, 3, 4, 5, 678900), '1900-01-02T03:04:05.678'),  # noqa: E501
+    (datetime(9999, 12, 31, 23, 59, 59, 999999), '9999-12-31T23:59:59.999'),  # noqa: E501
+    (datetime(1950, 5, 14, 20, 34, 52, 61000), '1950-05-14T20:34:52.061'),  # noqa: E501
+    (datetime(1999, 4, 3, 14, 7, 12), '1999-04-03T14:07:12.000'),  # noqa: E501
+    (datetime(2021, 1, 1, tzinfo=timezone.utc), '2021-01-01T00:00:00.000+00:00'),  # noqa: E501
+    (datetime(1, 1, 1, 1, 1, 1, 1, tzinfo=plus_1), '0001-01-01T01:01:01.000+01:00'),  # noqa: E501
+    (datetime(1900, 1, 2, 3, 4, 5, 678900, tzinfo=plus_12), '1900-01-02T03:04:05.678+12:00'),  # noqa: E501
+    (datetime(1999, 4, 3, 14, 7, 12, tzinfo=plus_0_15), '1999-04-03T14:07:12.000+00:15'),  # noqa: E501
+    (datetime(1950, 5, 14, 20, 34, 52, 61000, tzinfo=plus_12_30), '1950-05-14T20:34:52.061+12:30'),  # noqa: E501
+    (datetime(9999, 12, 31, 23, 59, 59, 999999, tzinfo=minus_1), '9999-12-31T23:59:59.999-01:00'),  # noqa: E501
+    (datetime(1950, 5, 14, 20, 34, 52, 61000, tzinfo=minus_12), '1950-05-14T20:34:52.061-12:00'),  # noqa: E501
+    (datetime(1999, 4, 3, 14, 7, 12, tzinfo=minus_0_15), '1999-04-03T14:07:12.000-00:15'),  # noqa: E501
+    (datetime(1950, 5, 14, 20, 34, 52, 61000, tzinfo=minus_12_30), '1950-05-14T20:34:52.061-12:30'),  # noqa: E501
+])
+def test_to_rfc3339(dt: datetime, expected: str):
+    assert to_rfc3339(dt) == expected
+
+
+@pytest.mark.parametrize("url_to_sanitize, expected", [
+    ('https://example.com', 'https://example.com'),
+    ('https://example.com/', 'https://example.com/'),
+    ('https://example.com/a/b/c', 'https://example.com/a/b/c'),
+    ('https://example.com/abc?xyz=123', 'https://example.com/abc'),
+    ('https://example.com/a/b/c', 'https://example.com/a/b/c'),
+    ('https://example.com/abc?xyz=123', 'https://example.com/abc'),
+    ('https://example.com/abc;x=1;y=2;z=3', 'https://example.com/abc'),
+    ('https://example.com:8000/abc?xyz=123', 'https://example.com:8000/abc'),
+    (b'https://example.com', b'https://example.com'),
+    (b'https://example.com/', b'https://example.com/'),
+    (b'https://example.com/a/b/c', b'https://example.com/a/b/c'),
+    (b'https://example.com/abc?xyz=123', b'https://example.com/abc'),
+    (b'https://example.com/a/b/c', b'https://example.com/a/b/c'),
+    (b'https://example.com/abc?xyz=123', b'https://example.com/abc'),
+    (b'https://example.com/abc;x=1;y=2;z=3', b'https://example.com/abc'),
+    (b'https://example.com:8000/abc?xyz=123', b'https://example.com:8000/abc'),
+
+    ('wss://example.com/abc?xyz=123', 'wss://example.com/abc'),
+    ('ftp://example.com/abc?xyz=123', 'ftp://example.com/abc'),
+    (b'wss://example.com/abc?xyz=123', b'wss://example.com/abc'),
+    (b'ftp://example.com/abc?xyz=123', b'ftp://example.com/abc'),
+
+    ('xyz', 'xyz'),
+    ('///', '/'),
+    ('/a/b/c', '/a/b/c'),
+    ('example.com/<<<<', 'example.com/<<<<'),
+    ('->example.com<-', '->example.com<-'),
+    (b'xyz', b'xyz'),
+    (b'///', b'/'),
+    (b'/a/b/c', b'/a/b/c'),
+    (b'example.com/<<<<', b'example.com/<<<<'),
+    (b'->example.com<-', b'->example.com<-'),
+
+    ('', None),
+    (b'', None),
+    ('  ', None),
+    (b'  ', None),
+    (None, None),
+    (123, None),
+    ([1, 2, 3], None),
+    ({'a': 1, b'b': 2, 'c': 3}, None),
+    (object(), None),
+    (lambda: 'example.com', None),
+])
+def test_sanitize_url(url_to_sanitize, expected):
+    assert sanitize_url(url_to_sanitize) == expected