+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 Changelog
 =========
 
+## 4.1.0 (2021-06-23)
+
+### Enhancements
+
+* Add support for breadcrumbs. These are short log statements that are attached to error reports to help diagnose what events led to the error. [See the docs for more information](https://docs.bugsnag.com//platforms/python/other/customizing-breadcrumbs/).
+  [#290](https://github.com/bugsnag/bugsnag-python/pull/290)
+
+### Bug fixes
+
+* Apply filtering with `params_filters` to bytes, not just strings
+  [Matt Bachmann](https://github.com/Bachmann1234)
+  [#267](https://github.com/bugsnag/bugsnag-python/pull/267)
+
 ## 4.0.3 (2021-04-28)
 
 ### Bug fixes